refactor: rename compliance command files
